projects/webimar/webimar-api/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap4==2024.10
Django==5.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-redis==5.4.0
djangorestframework==3.16.0
gunicorn==21.2.0
idna==3.10
lxml==5.4.0
numpy==2.2.5
psycopg2-binary==2.9.10
pykml==0.2.0
python-decouple==3.8
redis==5.0.1
requests==2.32.3
shapely==2.1.0
sqlparse==0.5.3
urllib3==2.4.0
djangorestframework-simplejwt
whitenoise==6.7.0
dj-database-url
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0

# Test dependencies (used by existing test modules)  
pytest==8.3.4
